Set content scrim and status bar scrim dynamically. Fix bug where wrong photo was being loaded and shown multiple times. Add callbacks to update Detail Activity UI.
Because the ViewPager loads multiple pages at once in order to cache additional fragments before they are visible, with the ImageView having been moved from the fragment to the activity, the same ImageView instance was being overwritten with the incorrect images. Code to set the image bitmap and the colors was refactored and moved to the activity level, to be triggered to be updated by a callback from the fragment.